feat: add styled-components to PreviousRecords
