+++ src/pages/PerviousRecords/components/RecordSelector/index.styled.js
@@ -1,0 +1,15 @@
+// styled-components
+import styled from "styled-components";
+// global components
+import { IonButton } from "@ionic/react";
+
+export const StyledIonButton = styled(IonButton)`
+  display: block;
+  margin: 15px auto;
+  width: 80%;
+  height: 60px;
+  font-weight: 700;
+  --background: #262424;
+  --box-shadow: none;
+  --border-radius: 16px;
+`;
