+++ src/pages/PerviousRecords/index.styled.js
@@ -1,0 +1,22 @@
+// styled-components
+import styled from "styled-components";
+// global components
+import { IonText } from "@ionic/react";
+
+export const StyledIonText = styled(IonText)`
+  font-size: 22px;
+  font-weight: bold;
+  color: white;
+`;
+
+export const PageButtons = styled.div`
+  text-align: center;
+  margin-top: 2%;
+  padding: 0% 14%;
+`;
+
+export const ResultContainer = styled.div`
+  color: white;
+  text-align: center;
+  padding: 0% 4%;
+`;
